Some style refactoring
