+++ item_warehouse_website/app/components/Warehouse/FullContentModal.client.tsx
@@ -10,7 +10,7 @@ import {
 } from "@mdi/js";
 import Icon from "@mdi/react";
 
-import styles from "styles/FullContentModal.module.scss";
+import styles from "styles/Warehouse/FullContentModal.module.scss";
 
 import { Button, Modal } from "react-bootstrap";
