+++ item_warehouse_website/app/styles/Settings/SettingsModal.module.scss
@@ -1,0 +1,28 @@
+@import "~bootstrap/scss/functions";
+@import "~bootstrap/scss/variables";
+
+.settings {
+  position: absolute;
+  bottom: 0;
+  width: inherit;
+  cursor: grabbing !important;
+} 
+
+
+.modalBody {
+  height: 75vh;
+}
+
+.button {
+  padding: 0 !important;
+  margin: 0.5em $spacer * 0.25;
+  width: $spacer * 2;
+  height: 1.9em;
+  border-width: 0 !important;
+  border-style: none !important;
+}
+
+.icon {
+  padding: 0;
+  margin: 0;
+}
