+++ item_warehouse_website/app/styles/Settings/WarehousePanel/WarehousePanel.module.scss
@@ -1,0 +1,5 @@
+.accordionHeader {
+    &:focus {
+        box-shadow: none !important;
+    }
+}
