+++ item_warehouse_website/app/styles/Warehouse/Cell.module.scss
