+++ item_warehouse_website/app/styles/Warehouse/FullContentModal.module.scss
